Leave low/medium difficulty zones early if beneficial
When a new planet appears it is sometimes beneficial to leave the current
planet early. An example follows.

The current zone is low difficulty and we have waited 10 seconds in the
current zone so far. Now, a new planet appears.

In the remaining 100 seconds (110 - 10), we would earn (100/110) * 2400
= 2181.8 in a high difficulty zone.

For the current zone, we must also account for the time spent so far. This
means we use the full score for it.

By leaving early we would get 2181.8 score over the time span covered by the
currently joined zone. That's an increase of 2182 - 600 = 1582.

So in this case it should be beneficial to switch zones, even though it
means the time spent in the current zone so far is wasted.